+++ src/Proton.cpp
@@ -1,6 +1,6 @@
 #include <ElectroSim/Proton.hpp>
 
-Proton::Proton(int x, int y, float radius) : Particle(x, y, radius) {
+Proton::Proton(float x, float y, float radius) : Particle(x, y, radius) {
     // Protons have a positive charge
     mCharge = CHARGE;
 }
